News Finance/requirements.txt
Django>=4.2.0
celery>=5.3.0
redis>=4.5.0
yfinance>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
whitenoise>=6.5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
djangorestframework>=3.14.0
lxml>=4.9.0
